rename 02_console to 01_console
