+++ ext/console/01_console.js
