+++ ext/console/lib.rs
@@ -1,7 +1,7 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 use std::path::PathBuf;
 
-deno_core::extension!(deno_console, esm = ["02_console.js"],);
+deno_core::extension!(deno_console, esm = ["01_console.js"],);
 
 pub fn get_declaration() -> PathBuf {
   PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("lib.deno_console.d.ts")
