+++ ext/web/01_dom_exception.js
@@ -21,7 +21,7 @@ const {
   SymbolFor,
 } = primordials;
 import * as webidl from "ext:deno_webidl/00_webidl.js";
-import { createFilteredInspectProxy } from "ext:deno_console/02_console.js";
+import { createFilteredInspectProxy } from "ext:deno_console/01_console.js";
 
 const _name = Symbol("name");
 const _message = Symbol("message");
