+++ ext/web/06_streams.js
@@ -78,7 +78,7 @@ const {
   WeakMapPrototypeHas,
   WeakMapPrototypeSet,
 } = primordials;
-import { createFilteredInspectProxy } from "ext:deno_console/02_console.js";
+import { createFilteredInspectProxy } from "ext:deno_console/01_console.js";
 import { assert, AssertionError } from "ext:deno_web/00_infra.js";
 
 /** @template T */
